All Auth Per Site Allow Registrations
Resolves #1. This will allow one to set the sign up visibility per site.
If it is toggled on, it should allow signups.
Else, not allow sign ups.
Makes the use of a default ALLOW_SIGN_UPS base parameter that can be
overwritten.

The class SiteSignupPolicy is based on SiteAwareModel in order to use
UUID's as these are not guessable.

Tests were also added to show the functionality.